ml/requirements.txt
hydra-core==1.1.1
facenet-pytorch==2.5.2
Pillow==8.4.0
torch==1.9.0
numpy==1.19.5
opencv-python==4.5.4.60
torchvision==0.10.0
omegaconf~=2.1.1
matplotlib~=3.5.1